refactor(data): modularize victory-config and close registry parity (#4072)
Extract civilian-build victory-config and topic-split GA params, register the
13 missing scalars, document Map tunables as out-of-scope, and centralize
region ids for in-package consumers.
